% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_min_saude.R
\name{baixa_min_saude}
\alias{baixa_min_saude}
\title{Baixa dados do ministério da saúde}
\usage{
baixa_min_saude()
}
\value{
Um data.frame com os dados do ministério da saúde de COVID 19
}
\description{
Função que baixa os dados do ministério da saúde
}
\examples{

}
