% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{capitais}
\alias{capitais}
\title{Base com codigos das capitais das unidades da federacao brasileira}
\format{
Uma tabela com 2 colunas e 27 linhas.
\describe{
  \item{capitais}{Variavel em caracteres com o nome das capitais das unidades da federacao brasileiras}
  \item{ibge}{Variavel numerica com o codigo ibge das capitais das unidades da federacao brasileiras}
}
}
\source{
Tabela Criada a partir dos dados do IBGE
}
\usage{
capitais
}
\description{
Uma base de dados com os codigos ibge das capitais das unidades da federacao brasileiras e brasilia
}
\keyword{datasets}
