% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifica-mortalidade-letalidade.R
\name{classifica_mortalidade_letalidade}
\alias{classifica_mortalidade_letalidade}
\title{Classificador dos estados}
\usage{
classifica_mortalidade_letalidade(dados_mor_let, produto_dt = FALSE)
}
\arguments{
\item{dados_mor_let}{base dados no formato mortalidade letalidade}

\item{produto_dt}{variavel logica que define se o produto da funcao e um data.table (se produto_dt == TRUE) ou um data.frame (se produto_dt = FALSE), padrao utilizado e FALSE}
}
\value{
Um data.frame com os dados do ministério da saúde de COVID 19
}
\description{
Funcao que recebe os dados de mortalidade e letalidade no formato do boletim e gera a classificao dos estados em diferentes informacoes a cerca da pandemia do COVID-19
}
\examples{

}
