% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desenha-grafs-semana.R
\name{desenha_grafs_semana}
\alias{desenha_grafs_semana}
\title{Desenha graficos da semana com dados semanais}
\usage{
desenha_grafs_semana(dados_semana)
}
\arguments{
\item{dados_semana}{Tabela de dados no formato data.frame utilizada para gerar os graficos em ggplot2 dos casos por semana}
}
\value{
Um data.frame com os dados do ministério da saúde de COVID 19 agrupados por semana
}
\description{
Funcao que recebe os dados casos por semana e gera os graficos utilizados no boletim em uma lista
}
\examples{

}
