% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desenha-series-metodo.R
\name{desenha_series_metodo}
\alias{desenha_series_metodo}
\title{Desenha grafico da comparacao entre metodologias}
\usage{
desenha_series_metodo(dados_ms, dados_ssp)
}
\arguments{
\item{lista_arquivos}{lista com data.frames com os dados de casos distrito federal limpos}

\item{lista_data}{lista com as datas de extracao dos data.frames da lista_arquivos}
}
\value{
Dois graficos comparando a serie de casos e obitos, em media movel de sete dias, para diferentes metodologias
}
\description{

}
\examples{

}
