% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desenha-series-ra.R
\name{desenha_series_ra}
\alias{desenha_series_ra}
\title{Desenha grafico das RAs por semana}
\usage{
desenha_series_ra(dados)
}
\arguments{
\item{dados}{de cadastro por semana de obitos e casos do Distrito Federal e Regioes Adminsitrativas}
}
\value{
lista com quatro graficos de casos e obitos, absolutos e por densidade para as RAs do Distrito Federal usando os dados por semana
}
\description{
Funcao que gera os graficos de casos e obitos, absolutos e por densidade para as RAs do Distrito Federal usando os dados por semana
}
\examples{

}
