% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desenha-top-ras.R
\name{desenha_top_ras}
\alias{desenha_top_ras}
\title{Desenha grafico das RAs com maior crescimento por semana}
\usage{
desenha_top_ras(dados_top)
}
\arguments{
\item{dados_top}{dados de cadastro por semana de obitos e casos das regioes administrativas com maior crescimento de casos e obitos nas ultimas duas semanas}
}
\value{
lista com dois graficos de casos e obitos absolutos e por densidade para as RAs do Distrito Federal com maior crescimento
}
\description{
Funcao que gera os graficos de casos e obitos, absolutos das últimas quatro semanas para as RAs que apresentaram maior crescimento de obitos e casos
}
\examples{

}
