% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-boletim.R
\name{gera_boletim}
\alias{gera_boletim}
\title{Gera os produtos utilizados no boletim COVID 19 da codeplan}
\usage{
gera_boletim(
  arquivo,
  local_resultados = "produto/",
  produto_dt = FALSE,
  max_data = NULL,
  prefixo = NULL
)
}
\arguments{
\item{arquivo}{especifica o arquivo com os dados do ministerios da saude}

\item{max_data}{especifica a data em que se avaliará os dados}
}
\value{
Um arquivo com os dados do ministério da saúde de COVID 19
}
\description{
Função que baixa os dados do ministério da saúde
}
\examples{

}
