% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-boletim-ra.R
\name{gera_boletim_ra}
\alias{gera_boletim_ra}
\title{Gera os produtos utilizados no boletim COVID 19 da codeplan}
\usage{
gera_boletim_ra(
  arquivo,
  local_resultados = "produto/",
  produto_dt = FALSE,
  max_data = NULL,
  prefixo = NULL
)
}
\arguments{
\item{arquivo}{caminho com os dados da SSP de cadastro em formato csv2}

\item{produto_dt}{variavel logica que define se o produto da funcao e um data.table (se produto_dt == TRUE) ou um data.frame (se produto_dt = FALSE), padrao utilizado e FALSE}
}
\value{
Um arquivo com os dados do ministério da saúde de COVID 19
}
\description{
Função que baixa os dados do ministério da saúde
}
\examples{

}
