% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-boletim-uf.R
\name{gera_boletim_uf}
\alias{gera_boletim_uf}
\title{Gera os produtos utilizados no boletim COVID 19 da codeplan}
\usage{
gera_boletim_uf(
  arquivo_uf,
  local_resultados = "produto/",
  produto_dt = FALSE,
  max_data = NULL,
  prefixo = NULL
)
}
\arguments{
\item{arquivo_uf}{Arquivo csv2 com os dados de UF do ministerios da Saude}

\item{local_resultados}{Diretorio em que os arquivos gerados serao salvos}

\item{produto_dt}{Valor logico se os produtos gerados internamente da funcao serao no formato data.table}

\item{max_data}{Data maxima em que serao avaliados os dados de covid}

\item{prefixo}{Qual o prefixo sera utilizado para os resultados, padrao e utilizar a data maxima dos arquivos avaliados}
}
\value{
Varios arquivos com o produto utilizados para fazer o boletim
}
\description{
Função que gera os dados do boletim a cerca sobre covid das unidades da federacao
}
\examples{

}
