% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-cadastro-semana.R
\name{gera_cadastro_semana}
\alias{gera_cadastro_semana}
\title{Gera dados de cadastro por semana}
\usage{
gera_cadastro_semana(
  dados_cadastro,
  data_inicio = "2020-01-01",
  produto_dt = FALSE
)
}
\arguments{
\item{dados_cadastro}{dados da SSP de cadastro por dia por localidade}

\item{data_inicio}{data em que se comecara a contagem das semanas}

\item{produto_dt}{variavel logica que define se o produto da funcao e um data.table (se produto_dt == TRUE) ou um data.frame (se produto_dt = FALSE), padrao utilizado e FALSE}
}
\value{
Um data.frame com os dados de serie temporal por semana dos dados de cadastro
}
\description{
Funcao que recebe os dados de casos e obitos por dia por regiao administrativa por data do cadastro da SSP e gera dados por semana
}
