% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-comparacao-series.R
\name{gera_comparacao_series}
\alias{gera_comparacao_series}
\title{Gera comparação de séries da SSP}
\usage{
gera_comparacao_series()
}
\value{

}
\description{
Função que faz a comparação das séries de dados da SSP
}
\examples{

}
