% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-produto.R
\name{gera_produto}
\alias{gera_produto}
\title{Gera os produtos utilizados no boletim COVID 19 da codeplan}
\usage{
gera_produto(
  local = "dados/",
  local_resultados = "produto/",
  produto_dt = FALSE,
  data_dados = NULL,
  prefixo = NULL
)
}
\arguments{
\item{local}{local para salvar a base da dados obtida do ministerio da saude, padrao e salvar para a pasta "dados" do diretorio atual}
}
\value{
Um arquivo com os dados do ministério da saúde de COVID 19
}
\description{
Função que baixa os dados do ministério da saúde
}
\examples{

}
