% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-serie-estadual.R
\name{gera_serie_estadual}
\alias{gera_serie_estadual}
\title{Gera serie estadual}
\usage{
gera_serie_estadual(dados_uf, produto_dt = FALSE, estados_selecionados = "DF")
}
\arguments{
\item{dados_uf}{Dados do ministerio da saude por UF limpos}

\item{produto_dt}{variavel logica que define se o produto da funcao e um data.table (se produto_dt == TRUE) ou um data.frame (se produto_dt = FALSE), padrao utilizado e FALSE}

\item{estados_selecionados}{Vetor de caracteres com os estados que seram selecionados do conjunto de dados estaduais}
}
\value{
Um data.frame com a serie historica de casos, obitos, casos acumulados e obitos para cada dia disponivel
}
\description{
Função que baixa os dados do ministério da saúde
}
\examples{

}
