% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-top-cadastro.R
\name{gera_top_cadastro}
\alias{gera_top_cadastro}
\title{Filtra localidades com maior crescimento dos dados de cadastro}
\usage{
gera_top_cadastro(dados_cadastro_semana, top = 6, produto_dt = FALSE)
}
\arguments{
\item{dados_cadastro_semana}{dados da SSP de cadastro por semana por localidade}

\item{top}{numero de regioes administrativas que seram filtradas}

\item{produto_dt}{variavel logica que define se o produto da funcao e um data.table (se produto_dt == TRUE) ou um data.frame (se produto_dt = FALSE), padrao utilizado e FALSE}
}
\value{
Um data.frame com os dados de serie temporal dos dados de cadastro das localidades com maior crescimento de casos e obitos
}
\description{
Funcao que recebe os dados de casos e obitos por semana por regiao administrativa por data do cadastro da SSP e filtra as observacoes com os principais casos
}
