% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le-min-saude.R
\name{le_min_saude}
\alias{le_min_saude}
\title{Le dados do ministério da saúde}
\usage{
le_min_saude(arquivo, max_data = NULL)
}
\arguments{
\item{arquivo}{especifica o arquivo com os dados do ministerios da saude}

\item{max_data}{especifica a data em que se avaliará os dados}
}
\value{
Um data.frame com os dados do ministério da saúde de COVID 19
}
\description{
Le o ultimo dado do ministerio da saude
}
\examples{

}
