% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpa-cadastro.R
\name{limpa_cadastro}
\alias{limpa_cadastro}
\title{Limpa dados de cadastro}
\usage{
limpa_cadastro(caminho_cadastro, produto_dt = FALSE)
}
\arguments{
\item{caminho_cadastro}{caminho com os dados da SSP de cadastro}
}
\value{
Um data.frame com os dados de serie temporal dos dados de cadastro
}
\description{
Funcao que recebe os dados de casos e obitos por observacao por regiao administrativa por data do cadastro da SSP e gera dados por dias
}
