% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpa-capitais.R
\name{limpa_capitais}
\alias{limpa_capitais}
\title{Limpa dados COVID-19 das capitais brasileiras}
\usage{
limpa_capitais(dados, produto_dt = FALSE)
}
\arguments{
\item{dados}{base dados carregada do ministerio da saude a ser lida}

\item{produto_dt}{variavel logica que define se o produto da funcao e um data.table (se produto_dt == TRUE) ou um data.frame (se produto_dt = FALSE), padrao utilizado e FALSE}
}
\value{
Um data.frame com os dados do ministério da saúde de COVID 19 para as capitais brasileiras
}
\description{
Recebe os dados no formato utilizado pelo ministerio da saude e limpa eles para o formato usado pelo boletim da codeplan na planilha de letalidade mortalidade das capitais brasileiras
}
\examples{

}
