% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpa-df.R
\name{limpa_df}
\alias{limpa_df}
\title{Gera serie DF extracao direta}
\usage{
limpa_df(dados_df, ra = FALSE, produto_dt = FALSE)
}
\arguments{
\item{dados_df}{data.frame com os dados de cada caso do Distrito Federal}

\item{ra}{Valor logico, se TRUE considera a serie historica por RA, se FALSE apenas para o Distrito Federal}

\item{produto_dt}{variavel logica que define se o produto da funcao e um data.table (se produto_dt == TRUE) ou um data.frame (se produto_dt = FALSE), padrao utilizado e FALSE}
}
\value{
Um data.frame com os dados da serie historica do Distrito Federal
}
\description{
Recebe os dados no formato utilizado pelo ministerio da saude e limpa eles para o formato usado pelo boletim da codeplan
}
