% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{pop}
\alias{pop}
\title{Base com populacao por estado}
\format{
Uma tabela com 2 colunas e 35 linhas.
\describe{
  \item{estado}{Variavel em caractere com o estado da localidade}
  \item{pop}{Variavel numerica com o numero de pessoas na localidade segundo o TCU para o ano de 2019}
}
}
\source{
Tabela Criada a partir dos dados do TCU
}
\usage{
pop
}
\description{
Uma base de dados da populacao por estados
}
\keyword{datasets}
