% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retira-valor.R
\name{retira_valor}
\alias{retira_valor}
\title{Retira valor}
\usage{
retira_valor(
  dados,
  variavel_condicional,
  condicao,
  variavel_retirada,
  percentual = FALSE
)
}
\arguments{
\item{dados}{data.frame a ser avaliado}

\item{variavel_condicional}{Variavel em que se colocora uma condicao logica de igualdade}

\item{condicao}{Valor no qual a variavel condicional deve ser igualada}

\item{variavel_retirada}{Coluna do qual se retirara o valor}

\item{percentual}{Se utilizara percentual no resultado}
}
\value{
Um valor numerico formatado
}
\description{
Retira o valor
}
\examples{

}
