% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retorna-dt-df.R
\name{retorna_dt_df}
\alias{retorna_dt_df}
\title{Retorna data.frame ou data.table}
\usage{
retorna_dt_df(dados, produto_dt)
}
\arguments{
\item{dados}{Um objeto da classe data.table}

\item{produto_dt}{variavel logica que define se o produto da funcao e um data.table (se produto_dt == TRUE) ou um data.frame (se produto_dt = FALSE), padrao utilizado e FALSE}
}
\value{
Um data.frame ou data.table
}
\description{
Recebe um data.table e determina se transforma em data.frame ou continua dt
}
\examples{
exemplo_mtcars <- data.table::data.table(mtcars)

retorna_dt_df(exemplo_mtcars, produto_dt = TRUE)

retorna_dt_df(exemplo_mtcars, produto_dt = FALSE)
}
