% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste-df.R
\docType{data}
\name{teste_df}
\alias{teste_df}
\title{Base de teste para o pacote}
\format{
Uma tabela com 17 colunas e 1300 linhas.
\describe{
  \item{id}{Variavel numerica com o id de cada observacao}
  \item{primeiros_sintomas}{Variavel em data com a data dos primeiros sintomas}
  \item{dataobito}{Variavel em data com a data do obito}
  \item{estado_de_saude}{Variavel em caractere com o estado de saude da pessoa}
}
}
\source{
Tabela Ficticia para testar as funcoes para gerar dados da serie do Distrito Federal
}
\usage{
teste_df
}
\description{
Uma base de dados ficticia para testar as funcoes do pacote do Distrito Federal
}
\keyword{datasets}
