% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desenha-crescimento-semana.R
\name{desenha_crescimento_semana}
\alias{desenha_crescimento_semana}
\title{Desenha grafico do crescimento da semana dos casos por UF}
\usage{
desenha_crescimento_semana(dados_uf)
}
\arguments{
\item{dados_uf}{Dados do ministerio da saude por UF limpos}
}
\value{
Um grafico de crescimento semanal usado no boletim semanal da codeplan
}
\description{
Recebe os dados de casos por UF no formato utilizado pelo ministerio da saude e gera um grafico de crescimento do numero de casos na ultima semana
}
\examples{

}
