% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-cadastro-semana.R
\name{gera_cadastro_semana}
\alias{gera_cadastro_semana}
\title{Gera dados de cadastro por semana}
\usage{
gera_cadastro_semana(
  dados_cadastro,
  data_inicio = "2012-01-01",
  produto_dt = FALSE
)
}
\arguments{
\item{caminho_cadastro}{caminho com os dados da SSP de cadastro}
}
\value{
Um data.frame com os dados de serie temporal por semana dos dados de cadastro
}
\description{
Funcao que recebe os dados de casos e obitos por dia por regiao administrativa por data do cadastro da SSP e gera dados por semana
}
