% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-produto.R
\name{gera_produto}
\alias{gera_produto}
\title{Gera os produtos utilizados no boletim COVID 19 da codeplan}
\usage{
gera_produto(dados, produto_dt)
}
\arguments{
\item{dados}{base dados carregada do ministerio da saude a ser lida}

\item{produto_dt}{variavel logica que define se o produto da funcao e um data.table (se produto_dt == TRUE) ou um data.frame (se produto_dt = FALSE), padrao utilizado e FALSE}
}
\value{
Uma lista com os dados do ministerio da saude no formato UF, os dados de mortalidade e letalidade do ministerio da saude e os dados de classificacao dos estados do ministerio da saude
}
\description{
Função que baixa os dados do ministério da saúde
}
\examples{

}
