% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le-min-saude.R
\name{le_min_saude}
\alias{le_min_saude}
\title{Le dados do ministério da saúde}
\usage{
le_min_saude(local = "dados/", data_dados = NULL)
}
\arguments{
\item{local}{especifica local do qual carregar a base da dados obtida do ministerio da saude, padrao e carregar a partir da pasta "dados" do diretorio atual}

\item{data_dados}{especifica de que data os dados devem ser carregados, o padrao e ler a ultima base disponivel}
}
\value{
Um data.frame com os dados do ministério da saúde de COVID 19
}
\description{
Le o ultimo dado do ministerio da saude
}
\examples{

}
