% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpa-cadastro.R
\name{limpa_cadastro}
\alias{limpa_cadastro}
\title{Limpa dados de cadastro}
\usage{
limpa_cadastro(arquivo, produto_dt = FALSE)
}
\arguments{
\item{arquivo}{caminho com os dados da SSP de cadastro em formato csv2}

\item{produto_dt}{variavel logica que define se o produto da funcao e um data.table (se produto_dt == TRUE) ou um data.frame (se produto_dt = FALSE), padrao utilizado e FALSE}
}
\value{
Um data.frame com os dados de serie temporal dos dados de cadastro
}
\description{
Funcao que recebe os dados de casos e obitos por observacao por regiao administrativa por data do cadastro da SSP e gera dados por dias
}
