% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_ra.R
\docType{data}
\name{pop_ra}
\alias{pop_ra}
\title{Base com populacao por regiao administrativa}
\format{
Uma tabela com 2 colunas e 35 linhas.
\describe{
  \item{estado}{Variavel em caracteres com o regiao administrativa da localidade}
  \item{pop}{Variavel numerica com o numero de pessoas na regiao administrativa}
}
}
\source{
Tabela Criada a partir dos dados da CODEPLAN
}
\usage{
pop_ra
}
\description{
Uma base de dados da populacao por regiao administrativa do Distrito Federal
}
\keyword{datasets}
